requirements.txt
asgiref==3.8.1
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
joblib==1.4.2
nltk==3.9.1
numpy==2.2.1
pdfminer.six==20240706
pycparser==2.22
PyJWT==2.10.1
regex==2024.11.6
scikit-learn==1.6.0
scipy==1.14.1
sqlparse==0.5.3
threadpoolctl==3.5.0
tqdm==4.67.1
